+++ module-array/src/main/java/ch/hsr/adv/ui/array/logic/ArrayLayouter.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.array.logic;
 
 import ch.hsr.adv.commons.array.logic.ConstantsArray;
+import ch.hsr.adv.commons.core.logic.domain.Module;
 import ch.hsr.adv.ui.core.logic.Layouter;
-import ch.hsr.adv.ui.core.logic.domain.Module;
 import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
