+++ module-array/src/main/java/ch/hsr/adv/ui/array/logic/domain/ArrayElement.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.array.logic.domain;
 
+import ch.hsr.adv.commons.core.logic.domain.ADVElement;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.commons.core.logic.domain.styles.presets.ADVDefaultStyle;
-import ch.hsr.adv.ui.core.logic.domain.ADVElement;
 
 /**
  * Represents one Item in a Array
@@ -16,7 +16,7 @@ public class ArrayElement implements ADVElement<String> {
     private ADVStyle style = new ADVDefaultStyle();
 
     @Override
-    public long getElementId() {
+    public long getId() {
         return id;
     }
