+++ module-array/src/test/java/ch/hsr/adv/ui/array/logic/ArrayParserTest.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.array.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ADVElement;
 import ch.hsr.adv.ui.array.logic.domain.ArrayElement;
 import ch.hsr.adv.ui.core.access.FileDatastoreAccess;
-import ch.hsr.adv.ui.core.logic.domain.ADVElement;
 import ch.hsr.adv.ui.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.ui.core.logic.util.ADVParseException;
 import com.google.gson.Gson;
